% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddSnippets.R
\name{addPackageSnippets}
\alias{addPackageSnippets}
\title{Export snippets}
\usage{
addPackageSnippets()
}
\value{
boolean invisible(FALSE) if nothing was added, invisible(TRUE) if snipped definitions were added
}
\description{
\code{addPackageSnippets} copies all (missing) snippet definitions
  in 'inst/rstudio/Rsnippets.txt' and 'Rmdsnippets.txt' to the RStudios user snippet location.
}
\examples{
\dontrun{addPackageSnippets()}
}
