% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bns2_pkg_data}
\alias{bns2_pkg_data}
\title{This is data included in the chcRne package}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50 rows and 28 columns.
}
\source{
{Data created in the project for function examples}
}
\usage{
bns2_pkg_data
}
\description{
This is data included in the chcRne package
}
\examples{
bns2_pkg_data
}
\author{
CHC
}
\keyword{data}
