% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chcRne-package.R
\docType{package}
\name{chcRne-package}
\alias{chcRne}
\alias{chcRne-package}
\title{chcRne: Functions and code snippets for the R&E team at the CHC}
\description{
If you are new to chcRne or are blanking on function names you can find documentation on the functions including examples here.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://smoorad99.github.io/chcRnePackage/}
}

}
\author{
\strong{Maintainer}: Saul Mooradian \email{slmooradian@csuchico.edu}

}
\keyword{internal}
