% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_and_percent.R
\name{count_and_percent}
\alias{count_and_percent}
\title{Count and Percent}
\usage{
count_and_percent(var, ...)
}
\arguments{
\item{var}{The variable for which the count and percent are calculated.}

\item{...}{One or more categories of `var` to count and calculate percentages.}
}
\value{
A string including the total count and percentage of the
categories in the input variable.
}
\description{
Count and Percent
}
\examples{
count_and_percent(var = exp.data$no_goingout_cooking, "1")

}
