% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp.data}
\alias{exp.data}
\title{This is data included in my package}
\format{
An object of class \code{data.frame} with 20 rows and 3 columns.
}
\source{
{Data created in the project for function examples}
}
\usage{
exp.data
}
\description{
This is data included in my package
}
\examples{
exp.data
}
\author{
Saul Mooradian
}
\keyword{data}
