% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dupes.R
\name{find_dupes}
\alias{find_dupes}
\title{Finding dupes in pre/post data}
\usage{
find_dupes(data, vars)
}
\arguments{
\item{data}{A data frame containing at least one key ID variable used to identify any duplicates}

\item{vars}{The ID variable in quotes. If more than one variable is needed to identify unique records, pass a comma separated character vector e.g. `c("a", "b")`}
}
\value{
A data set containing the ID variables and number of copies found
}
\description{
Finding dupes in pre/post data
}
\examples{
suppressMessages(library(dplyr))
testdata <- datasets::sleep \%>\% rbind(c(.4, 1, 1)) \%>\% filter(extra != 3.7) # for example purposes
find_dupes(testdata, "ID")
# For more than one key variable
find_dupes(testdata, c("ID", "group"))
}
