% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_perct.R
\name{get_perct}
\alias{get_perct}
\title{Calculate the percentage of occurrences for a value in a column}
\usage{
get_perct(x, value)
}
\arguments{
\item{x}{A number or character that contains the value of interest.}

\item{value}{The specific value to search for in the column `x`.}
}
\value{
A string that provides the count of the specific value, the total count
of non-NA values, and the percentage of the specific value in the format of "count/total (percentage%)".
}
\description{
This function calculates the percentage of a value within a column,
excluding NA values. It returns the count of the specific value,
the total count of non-NA values, and the percentage of the specific value out of all the
non-NA values.
}
\examples{
get_perct(bns2_pkg_data$q14_9, "Yes")

}
