% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_n_reporting.R
\name{print_n_reporting}
\alias{print_n_reporting}
\title{Number and percent of non-NA entries in a column}
\usage{
print_n_reporting(data, x)
}
\arguments{
\item{data}{A dataframe containing the dataset to be analyzed.}

\item{x}{The name of the column of interest.}
}
\value{
A string formatted as "(n=X, Y of Z reporting)", where X is
the number of non-NA entries in the specified column, Y is the percentage of
non-NA entries out of the total number of rows in the dataframe, and Z is the
total number of rows in the dataframe.
}
\description{
This function calculates and returns the number and
percentage of non-NA entries in a column of a dataframe.
}
\examples{
print_n_reporting(bns2_pkg_data, "q14_20")

}
