% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_table.R
\name{question_table}
\alias{question_table}
\title{Table of Counts Percentages for Single Question}
\usage{
question_table(data, x, cnames, kbl_styling = TRUE)
}
\arguments{
\item{data}{A data frame containing the dataset to be analyzed.}

\item{x}{The name of the column in `data` containing the question}

\item{cnames}{A character vector of length 2 specifying the column names for the
resulting table. The first element is for the category name and the second for the percentages.}

\item{kbl_styling}{A logical value indicating whether to style the table with `kableExtra`.
Defaults to TRUE. If TRUE, the table is returned with `kableExtra` styling applied.}
}
\value{
A table of percentages for the specified question. If `kbl_styling` is TRUE,
  the table is styled with `kableExtra` and suitable for markdown rendering.
  Otherwise, a plain data frame is returned.
}
\description{
This function takes a dataset and a specified column representing a single
categorical or binary variable (e.g., level of education) and creates a table of
response percentages. The table can be optionally styled using `kableExtra`
for better visualization.
}
\examples{
question_table(data = bns2_pkg_data, x = 'q13', cnames = c("Level of Education", "Yes \%"), kbl_styling = TRUE)

}
