% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.na.levels.R
\name{remove.na.levels}
\alias{remove.na.levels}
\title{Remove NA levels in factor variables}
\usage{
remove.na.levels(x)
}
\arguments{
\item{x}{A factor variable containing NA as a level.}
}
\value{
The same factor variable with NA levels explicitly marked as NA.
}
\description{
`remove.na.levels()` processes a factor variable that includes NA as a level,
removing the level while preserving the NA responses.
}
\examples{
# As seen in the table below, this factor variable has `NA` as a level.

table(bns2_pkg_data$q13.na)

# Applying `remove.na.levels` TO remove the `NA` level.

remove.na.levels(bns2_pkg_data$q13.na) |>
  table()

# We can use the `across` function from `dplyr` to apply this function to multiple variables in of fell swoop.

}
