% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_binary.R
\name{to_binary}
\alias{to_binary}
\title{Convert Yes/No variables to binary}
\usage{
to_binary(data, these.cols, prefix = FALSE)
}
\arguments{
\item{data}{A dataframe containing the variables to be converted.}

\item{these.cols}{A prefix used to select all columns with the prefix or the columns specified like you would using dplyr::select()}

\item{prefix}{Whether or not you are using a prefix to select columns. This argument defaults to false
i.e. assumes your not using a prefix unless you tell it otherwise.}
}
\value{
The original dataframe with specified variables converted
to binary.
}
\description{
This function converts all "String"/"NA" responses in the selected columns to
binary (1/0) where a replaces the strings and a 0 replaces the NAs.
}
\examples{
# Convert "Yes"/"No" responses in columns starting with "q14" to 1/0
df_converted <- to_binary(data = bns2_pkg_data, these.cols = "q14_", prefix = TRUE)

# View the converted dataframe side-by-side
old <- bns2_pkg_data |> dplyr::select(q14_1, q14_4)
new <- df_converted |> dplyr::select(q14_1, q14_4)
cbind(old, new)

}
