% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y.off.R
\name{y.off}
\alias{y.off}
\title{Function for plot_frq y.offset}
\usage{
y.off(datavar)
}
\arguments{
\item{datavar}{Data frame $ variable}
}
\value{
A `plot_frq` bar chart with counts and percents on screen.
}
\description{
The purpose of `y.off` is to ensure the count and percent remains "on screen" regardless of number of rows when using `plot_frq()`.
}
\examples{
df <- bns2_pkg_data |> head(15)
sjPlot::plot_frq(df$q13)
sjPlot::plot_frq(df$q13, y.offset = y.off(bns2_pkg_data$q13))
}
