% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yes_pct.R
\name{yes_pct}
\alias{yes_pct}
\title{Yes and percent}
\usage{
yes_pct(df, var1, var2, value, plot = FALSE)
}
\arguments{
\item{df}{A data frame containing the data set to be analyzed.}

\item{var1}{The variable containing the levels to be grouped and summarized.}

\item{var2}{The variable you want to count the number of occurrences of a certain value.}

\item{value}{The value from var2 to be counted.}

\item{plot}{If plot = FALSE `yes_pct` will return a data
frame with summarized data. If plot = TRUE `yes_pct` will return a ggplot object
to visualize the summarized data. Defaults to plot = FALSE.}
}
\description{
yes_pct calculates the count and percent of a user specified value in var2 for each level of a var1.
}
\examples{
yes_pct(bns2_pkg_data, q13, q14_1, value = "Yes")
yes_pct(bns2_pkg_data, q13, q14_1, value = "Yes", plot = TRUE)
}
