% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_and_percent.R
\name{count_and_percent}
\alias{count_and_percent}
\title{Creates a string with count and percent}
\usage{
count_and_percent(var, ...)
}
\arguments{
\item{var}{A variable for which the counts and percents are calculated.}

\item{...}{One or more categories of `var` to get the counts and calculate the percentages. If the ellipsis is left blank, the function
will calculate the count and percent for the category with highest count.}
}
\value{
A string including the total count and percentage of the
categories in the input variable.
}
\description{
This function calculates counts and percentages for specified categories within a variable
then returns a string including the total count and total percent of those categories relative to the
variable they reside in.
}
\examples{
count_and_percent(var = bns2_pkg_data$q14_1, "Yes")

}
