% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nperc_tbl_MATA.R
\name{nperc_tbl_MATA}
\alias{nperc_tbl_MATA}
\title{Creates frequency table}
\usage{
nperc_tbl_MATA(df, vars, value, row.names, punc)
}
\arguments{
\item{df}{A data frame used to create the table.}

\item{vars}{A character vector of variable names to include in the table (all vars have to be binary).}

\item{value}{The value in the rows of `vars` (only one value across all variables).}

\item{row.names}{The name of each row in the table. Should be equivalent to each option in the select all that apply question}

\item{punc}{The punctuation at the end of the row names.}
}
\value{
A frequency table
}
\description{
This function creates a table showing the n counts and percentages for each option in a mark all that apply question.
}
\examples{
# Save the variable names to questions and the names of the questions to rnames
cols <- c("q14_21", "q14_22", "q14_23")

nperc_tbl_MATA(df = bns2_pkg_data,
               vars = cols,
               value = "Yes",
               row.names = c("I am a broke college student", "I like cooking", "No good restaurants"),
               punc = ".")


}
