% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnta_unanswered_to_miss.R
\name{pnta.unanswered.to.miss}
\alias{pnta.unanswered.to.miss}
\title{PNTA to NA}
\usage{
pnta.unanswered.to.miss(data, prefix, pnta)
}
\arguments{
\item{data}{The data frame of interest.}

\item{prefix}{The prefix used to identify the relevant columns.}

\item{pnta}{The variable that contains the "Prefer not to answer" option.}
}
\value{
A modified data frame with NA assigned to cells in specified columns for rows with PNTA responses.
}
\description{
This function sets all options (vars) in a select all that apply question to NA when the "Prefer not to answer" option was selected
}
\examples{
pnta.unanswered.to.miss(data = exp.data,
                                  prefix = "no_going",
                                  pnta = exp.data$Q3)
}
