% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pnta.unanswered.to.miss}
\alias{pnta.unanswered.to.miss}
\title{PNTA to MISS}
\usage{
pnta.unanswered.to.miss(prefix, data, pnta)
}
\arguments{
\item{prefix}{A commonality/string shared by all sub-questions in the select all that applies question.}

\item{data}{The data frame of interest}

\item{pnta}{The variable that has the "Prefer not to answer" option}
}
\value{
A new data frame
}
\description{
Sets all options (vars) in select all that apply questions to NA when the "Prefer not to answer" option was selected
}
\examples{
pnta.unanswered.to.miss(prefix = "no_going", data = exp.data, pnta = exp.data$Q3)
}
