% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnta_unanswered_to_miss.R
\name{pnta.unanswered.to.miss}
\alias{pnta.unanswered.to.miss}
\title{PNTA to NA}
\usage{
pnta.unanswered.to.miss(data, prefix, pnta)
}
\arguments{
\item{data}{The data frame of interest.}

\item{prefix}{The prefix used to identify the relevant columns.}

\item{pnta}{The variable that contains the "Prefer not to answer" option.}
}
\value{
A modified data frame with NA assigned to cells in specified rows where prefer not to answer was selected.
}
\description{
This function sets all options (vars) in a select all that apply question to NA when the "Prefer not to answer" option was selected
}
\examples{
subset <- bns2_pkg_data[c(14:15, 17:18), ]

updated_data <- pnta.unanswered.to.miss(data = subset,
                                  prefix = "q14_2",
                                  pnta = subset$q14_30)
updated_data |> dplyr::select(q14_20:q14_26, q14_30)

}
