% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.na.levels.R
\name{remove.na.levels}
\alias{remove.na.levels}
\title{Remove NA Levels in Ordered Factors}
\usage{
remove.na.levels(x)
}
\arguments{
\item{x}{A factor variable containing}
}
\value{
The same factor variable with NA levels explicitly marked as `NA`.
}
\description{
Remove NA Levels in Ordered Factors
}
\examples{
# Given a factor with NA as one of its levels

}
