% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y.off.R
\name{y.off}
\alias{y.off}
\title{plot_frq y.offset function}
\usage{
y.off(datavar)
}
\arguments{
\item{datavar}{Data frame $ variable}
}
\value{
A `plot_frq` bar chart with counts and percents on screen.
}
\description{
The sole purpose of `y.off()` is to make sure the count and percent above each bar generated by the `plot_frq` remains "on screen".
}
\examples{
plot_frq(bns2_pkg_data$q13)
plot_frq(bns2_pkg_data$q13, y.offset = y.off(bns2_pkg_data$q13))



}
